.tools/web-terminal/requirements.txt
pyobjc-framework-Cocoa
pyobjc-framework-WebKit
